🎨 Moved Breadcrumb schema to breadcrumbs folder
