+++ components/blocks/index.ts
@@ -1,7 +1,5 @@
 import type { Template } from "tinacms";
-
 import { clientLogosBlockSchema } from "../../components/blocks/clientLogos";
-import { BreadcrumbSchema } from "../../tina/collections/breadcrumbs";
 import { bookingButtonSchema } from "../bookingButton/bookingButton";
 import { utilityButtonSchema } from "../button/utilityButton";
 import { clientListSchema } from "../company/clientList";
@@ -21,6 +19,7 @@ import { organizerSchema } from "../usergroup/organizer";
 import { customDownloadButtonSchema } from "./CustomDownloadButton";
 import { aboutUsBlockSchema } from "./aboutUs";
 import { agendaSchema } from "./agenda";
+import { BreadcrumbSchema } from "./breadcrumbs/breadcrumbs-schema";
 import { builtOnAzureBlockSchema } from "./builtOnAzure";
 import { carouselBlockSchema } from "./carousel";
 import { citationBlockSchema } from "./citation";
